Migrates substrate events and methods
