+++ utils/ethereum/testing.go
@@ -1,0 +1,2 @@
+package ethutils
+
