+++ utils/logs.go
@@ -1,7 +1,7 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package testutils
+package utils
 
 import (
 	log "github.com/ChainSafe/log15"
